src/requirements.txt
Pillow
transformers>=4.31.0
peft==0.5.0
accelerate==0.21.0
bitsandbytes==0.40.2
safetensors>=0.3.3